+++ src/OmniSharp.Abstractions/Models/v1/DiagnosticsRequest.cs
@@ -1,0 +1,15 @@
+using System;
+using OmniSharp.Mef;
+
+namespace OmniSharp.Models
+{
+    [OmniSharpEndpoint(OmnisharpEndpoints.Diagnostics, typeof(DiagnosticsRequest), typeof(DiagnosticsResponse))]
+    public class DiagnosticsRequest : Request
+    {
+    }
+
+    public class DiagnosticsResponse : IAggregateResponse
+    {
+        public IAggregateResponse Merge(IAggregateResponse response) { return response; }
+    }
+}
